creating log_parser
